bot/requirements.txt
python-telegram-bot==13.15
SQLAlchemy==1.4.52
psycopg2-binary==2.9.9